requirements.txt
langchain>=0.0.300
python-dotenv>=1.0.0
requests>=2.31.0
langchain-ollama>=0.1.3
